fix: fixes after search + boostrap cleanups
