+++ .gitignore
@@ -85,5 +85,4 @@ package-lock.json
 packages/cache/node_modules/*
 newrelic_agent.log
 /generated/
-.queue/events/queue.json
-.queue/events/state.json
+.queue/**
