+++ packages/core/src/modules/auth/lib/setup-app.ts
@@ -368,7 +368,9 @@ async function ensureDefaultRoleAcls(
       'attachments.view',
       'attachments.manage',
       'query_index.*',
+      'search.*',
       'vector.*',
+      'feature_toggles.*',
       'configs.system_status.view',
       'configs.cache.view',
       'configs.cache.manage',
